Offload update installation to the subsequent loader process
Previously, the local copy of the loader was in charge of installing the
files from the downloaded package. This is quite limiting, because if
the archive changes/has additional files, the loader will not know of
them.

By allowing the downloaded loader to handle the installation, we can
handle future circumstances where we may include additional files in the
downloaded package.