Compile and install packages per file rather than directory
This is quite a radical change for both the compiler, installer and
verification method in the kernel.

Previously, packages would represent entire subdirectories. Now,
they represent individual files. This is to minimise the chances of
a package surpassing 4GiB, thereby allowing SFX or ISO files to be
created for containing the installation packages.

This commit also improves the verbosity and consistency of console
information.