Introduce the new HXE kernel and SPV3 loader
This is a significant update that drastically improves the initiation
and configuration capabilities of the SPV3.2 loader. Most of the changes
are under the hood; however, the configuration screen will look quite
different as to reflect HXE's new abilities.

This is achieved by adapting the loader to the new HXE kernel, which
introduces a myriad of features and improvements, including:

-   a revamped kernel and configuration structure, fully documented
-   ability to optionally specify custom configuration _and_ executable
    objects, with sane defaults & detection
-   better granularity over the profile enhancement & initiation tweaks,
    for example: enabling custom resolution but disabling FPS uncap
-   new initiation tweaks, including gamma and post-processing disabling
-   improved support for HCE booting, SPV3.1 unlocking, and SPV3.2
    tweaking (including shaders and compatibility improvements)
-   graceful initiation of HCE/SPV3.2, for example, by killing running
    HCE processes and disabling EAX if the DSOAL file does not exist
-   a brand new configuration UI, which is invoked using hxe.exe -config

With the above features in mind, the SPV3.2 loader is now capable of
initiation and configuration features such as:

-   providing two video modes: one focusing on video stability by using
    HCE's native video mode, and the other focusig on FPS uncapping
-   specifying a custom resolution, refresh rate, gamma and display
-   toggling shaders, sound enhancement, and v-sync on or off

Before, HXE settings were stored in its own separate path. Due to the
ability to provide custom configuration objects, both the HXE kernel and
SPV3 loader's settings are stored in the same singular location:

    %APPDATA%\SPV3

Additionally, this update also introduces quality-of-life improvements
such as shader presets, and a friendlier configuration screen.

Due the ability to toggle post-processing and choosing presets, the
Ultra Low GFX mode has been discarded.

Documentation for both HXE and SPV3 have been updated, accordingly.