Declare initiation filename depending on OpenSauce's presence
This commit updates the declaration of the initiation file in the Paths
module of HXE. Previously, the initiation file was a constant initc.txt
value.

Now, it will be either initc.txt or init.txt depending on whether the
OpenSauce DLL is present, respectively.

Reported by github:VoidsShadow and resolves #20 on GitHub.