Introduce the HCE Progress abstract
This commit introduces the HCE Progress abstract and instructs the SPV3
Progress concrete to inherit it.

The abstract handles the loading of the Mission and Difficulty data from
the savegame.bin binary, and expects the inheriting concretes to handle
the representation of the binary data as Mission and Difficulty objects.