Add the Kernel Mode states for SPV3.2 & SPV3.3
This commit introduces the Kernel States in XML for SPV3.2 and SPV3.3.
That is, the XMl files are serialised versions of the Mode object with
the respective Mission, Difficulty and Configuration values for each
version of SPV3.

HXE will eventually use those files when handling Campaign resuming and
figuring out what configuration options it can & should apply in the
initiation file.