Replace version inference file and logic
This commit declares the Paths.Legacy (legacy.txt) file and discards the
Paths.Version (version.txt) file.

Previously, the presence of the Version file would represent the SPV3
installation being that of >=3.3. Now, the presence of the Legacy file
represents the installation being <=3.2.

The benefit of this commit is that loader will work with 3.3 immediately
though at the expense of players on legacy versions needing to create
the legacy.txt file in their directory.

Should breaking changes occur again in the future for the SPV3 project,
the legacy.txt file can be taught to contain the SPV3 version and thus
instruct the loader to adapt its backwards compatibility, accordingly.

(cherry picked from commit 12f092a0ff3c2ae475e3434412ad21b59b028386)