Implement the HXE SFX module
This module is the successor to the Installer one. It differs from the
Installer by creating a HXE executable with all of the installable data
compressed & embedded inside it.

This allows the creation of completely self-contained installers that
can be distributed as is, without the need for temporary extractions of
installation data.